+++ libraries/chain/include/graphene/chain/protocol/nft_ops.hpp
@@ -18,7 +18,7 @@ namespace graphene { namespace chain {
       std::string symbol;
       std::string base_uri;
       optional<account_id_type> revenue_partner;
-      optional<double> revenue_split;
+      optional<uint16_t> revenue_split;
       bool is_transferable = false;
       bool is_sellable = true;
 
@@ -38,7 +38,7 @@ namespace graphene { namespace chain {
       optional<std::string> symbol;
       optional<std::string> base_uri;
       optional<account_id_type> revenue_partner;
-      optional<double> revenue_split;
+      optional<uint16_t> revenue_split;
       optional<bool> is_transferable;
       optional<bool> is_sellable;
